Split DocSidebarItem by item type
